+++ app/src/main/java/kolskypavel/ardfmanager/ui/categories/CategoryFragment.kt
@@ -118,7 +118,7 @@ class CategoryFragment : Fragment() {
             R.id.category_menu_edit_race -> {
                 findNavController().navigate(
                     BottomNavDirections.modifyRaceProperties(
-                        RaceEditDialogFragment.RaceEditAcctions.EDIT,
+                        RaceEditDialogFragment.RaceEditActions.EDIT,
                         0,
                         selectedRaceViewModel.race.value
                     )